requirements-prod.txt
# Production requirements
psycopg2-binary==2.9.9
channels>=4.0
channels-redis>=4.0
daphne>=4.0
redis==5.0.1
celery==5.3.4
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
pandas==2.2.3
openpyxl==3.1.5
django-guardian==2.4.0
djangorestframework==3.14.0
django-filter==24.3

# Production dependencies
gunicorn>=22.0.0
whitenoise>=6.0.0
python-dotenv==1.0.0
